feat(python): allow internal api to get pointer to values buffer
