+++ py-polars/src/npy.rs
@@ -4,8 +4,15 @@ use ndarray::IntoDimension;
 use numpy::npyffi::types::npy_intp;
 use numpy::npyffi::{self, flags};
 use numpy::{Element, PyArray1, ToNpyDims, PY_ARRAY_API};
+use polars_core::prelude::*;
 use polars_core::utils::arrow::types::NativeType;
+use polars_core::with_match_physical_numeric_polars_type;
 use pyo3::prelude::*;
+use pyo3::types::PyTuple;
+
+use crate::error::PyPolarsErr;
+use crate::raise_err;
+use crate::series::PySeries;
 
 /// Create an empty numpy array arrows 64 byte alignment
 ///
@@ -51,5 +58,84 @@ pub fn get_refcnt<T>(pyarray: &PyArray1<T>) -> isize {
         refcnt - (2 << 60)
     } else {
         refcnt
+    }
+}
+
+macro_rules! impl_ufuncs {
+    ($name:ident, $type:ident, $unsafe_from_ptr_method:ident) => {
+        #[pymethods]
+        impl PySeries {
+            // applies a ufunc by accepting a lambda out: ufunc(*args, out=out)
+            // the out array is allocated in this method, send to Python and once the ufunc is applied
+            // ownership is taken by Rust again to prevent memory leak.
+            // if the ufunc fails, we first must take ownership back.
+            pub fn $name(&self, lambda: &PyAny) -> PyResult<PySeries> {
+                // numpy array object, and a *mut ptr
+                Python::with_gil(|py| {
+                    let size = self.len();
+                    let (out_array, av) =
+                        unsafe { aligned_array::<<$type as PolarsNumericType>::Native>(py, size) };
+
+                    debug_assert_eq!(get_refcnt(out_array), 1);
+                    // inserting it in a tuple increase the reference count by 1.
+                    let args = PyTuple::new(py, &[out_array]);
+                    debug_assert_eq!(get_refcnt(out_array), 2);
+
+                    // whatever the result, we must take the leaked memory ownership back
+                    let s = match lambda.call1(args) {
+                        Ok(_) => {
+                            // if this assert fails, the lambda has taken a reference to the object, so we must panic
+                            // args and the lambda return have a reference, making a total of 3
+                            assert_eq!(get_refcnt(out_array), 3);
+
+                            let validity = self.series.chunks()[0].validity().cloned();
+                            let ca = ChunkedArray::<$type>::new_from_owned_with_null_bitmap(
+                                self.name(),
+                                av,
+                                validity,
+                            );
+                            PySeries::new(ca.into_series())
+                        }
+                        Err(e) => {
+                            // return error information
+                            return Err(e);
+                        }
+                    };
+
+                    Ok(s)
+                })
+            }
+        }
+    };
+}
+impl_ufuncs!(apply_ufunc_f32, Float32Type, unsafe_from_ptr_f32);
+impl_ufuncs!(apply_ufunc_f64, Float64Type, unsafe_from_ptr_f64);
+impl_ufuncs!(apply_ufunc_u8, UInt8Type, unsafe_from_ptr_u8);
+impl_ufuncs!(apply_ufunc_u16, UInt16Type, unsafe_from_ptr_u16);
+impl_ufuncs!(apply_ufunc_u32, UInt32Type, unsafe_from_ptr_u32);
+impl_ufuncs!(apply_ufunc_u64, UInt64Type, unsafe_from_ptr_u64);
+impl_ufuncs!(apply_ufunc_i8, Int8Type, unsafe_from_ptr_i8);
+impl_ufuncs!(apply_ufunc_i16, Int16Type, unsafe_from_ptr_i16);
+impl_ufuncs!(apply_ufunc_i32, Int32Type, unsafe_from_ptr_i32);
+impl_ufuncs!(apply_ufunc_i64, Int64Type, unsafe_from_ptr_i64);
+
+fn get_ptr<T: PolarsNumericType>(ca: &ChunkedArray<T>) -> usize {
+    let arr = ca.downcast_iter().next().unwrap();
+    arr.values().as_ptr() as usize
+}
+
+#[pymethods]
+impl PySeries {
+    pub fn get_ptr(&self) -> PyResult<usize> {
+        let s = self.series.to_physical_repr();
+        let arrays = s.chunks();
+        if arrays.len() != 1 {
+            let msg = "Only can take pointer, if the 'series' contains a single chunk";
+            raise_err!(msg, ComputeError);
+        }
+        Ok(with_match_physical_numeric_polars_type!(s.dtype(), |$T| {
+            let ca: &ChunkedArray<$T> = s.as_ref().as_ref().as_ref();
+            get_ptr(ca)
+        }))
     }
 }
