+++ py-polars/src/series.rs
@@ -4,7 +4,7 @@ use polars_core::series::IsSorted;
 use polars_core::utils::{flatten_series, CustomIterTools};
 use pyo3::exceptions::{PyRuntimeError, PyValueError};
 use pyo3::prelude::*;
-use pyo3::types::{PyBytes, PyList, PyTuple};
+use pyo3::types::{PyBytes, PyList};
 use pyo3::Python;
 
 use crate::apply::series::{call_lambda_and_extract, ApplyLambda};
@@ -12,10 +12,10 @@ use crate::arrow_interop::to_rust::array_to_rust;
 use crate::dataframe::PyDataFrame;
 use crate::error::PyPolarsErr;
 use crate::list_construction::py_seq_to_list;
-use crate::npy::{aligned_array, get_refcnt};
 use crate::prelude::*;
 use crate::set::set_at_idx;
 use crate::{apply_method_all_arrow_series2, arrow_interop};
+
 #[pyclass]
 #[repr(transparent)]
 #[derive(Clone)]
@@ -1141,64 +1141,6 @@ impl PySeries {
         self.series.is_sorted(options)
     }
 }
-
-macro_rules! impl_ufuncs {
-    ($name:ident, $type:ident, $unsafe_from_ptr_method:ident) => {
-        #[pymethods]
-        impl PySeries {
-            // applies a ufunc by accepting a lambda out: ufunc(*args, out=out)
-            // the out array is allocated in this method, send to Python and once the ufunc is applied
-            // ownership is taken by Rust again to prevent memory leak.
-            // if the ufunc fails, we first must take ownership back.
-            pub fn $name(&self, lambda: &PyAny) -> PyResult<PySeries> {
-                // numpy array object, and a *mut ptr
-                Python::with_gil(|py| {
-                    let size = self.len();
-                    let (out_array, av) =
-                        unsafe { aligned_array::<<$type as PolarsNumericType>::Native>(py, size) };
-
-                    debug_assert_eq!(get_refcnt(out_array), 1);
-                    // inserting it in a tuple increase the reference count by 1.
-                    let args = PyTuple::new(py, &[out_array]);
-                    debug_assert_eq!(get_refcnt(out_array), 2);
-
-                    // whatever the result, we must take the leaked memory ownership back
-                    let s = match lambda.call1(args) {
-                        Ok(_) => {
-                            // if this assert fails, the lambda has taken a reference to the object, so we must panic
-                            // args and the lambda return have a reference, making a total of 3
-                            assert_eq!(get_refcnt(out_array), 3);
-
-                            let validity = self.series.chunks()[0].validity().cloned();
-                            let ca = ChunkedArray::<$type>::new_from_owned_with_null_bitmap(
-                                self.name(),
-                                av,
-                                validity,
-                            );
-                            PySeries::new(ca.into_series())
-                        }
-                        Err(e) => {
-                            // return error information
-                            return Err(e);
-                        }
-                    };
-
-                    Ok(s)
-                })
-            }
-        }
-    };
-}
-impl_ufuncs!(apply_ufunc_f32, Float32Type, unsafe_from_ptr_f32);
-impl_ufuncs!(apply_ufunc_f64, Float64Type, unsafe_from_ptr_f64);
-impl_ufuncs!(apply_ufunc_u8, UInt8Type, unsafe_from_ptr_u8);
-impl_ufuncs!(apply_ufunc_u16, UInt16Type, unsafe_from_ptr_u16);
-impl_ufuncs!(apply_ufunc_u32, UInt32Type, unsafe_from_ptr_u32);
-impl_ufuncs!(apply_ufunc_u64, UInt64Type, unsafe_from_ptr_u64);
-impl_ufuncs!(apply_ufunc_i8, Int8Type, unsafe_from_ptr_i8);
-impl_ufuncs!(apply_ufunc_i16, Int16Type, unsafe_from_ptr_i16);
-impl_ufuncs!(apply_ufunc_i32, Int32Type, unsafe_from_ptr_i32);
-impl_ufuncs!(apply_ufunc_i64, Int64Type, unsafe_from_ptr_i64);
 
 macro_rules! impl_set_with_mask {
     ($name:ident, $native:ty, $cast:ident, $variant:ident) => {
